lagchainFundamantels/pyproject.toml
[project]
name = "langchain-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.0.27",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
